pyproject.toml
[project]
name = "student-attendance-django"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.5",
    "gunicorn>=23.0.0",
    "mysql-connector-python>=9.4.0",
    "mysqlclient",
    "numpy>=2.3.2",
    "opencv-python>=4.11.0.86",
    "python-dotenv",
    "rich>=14.1.0",
    "scipy>=1.16.1",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "ultralytics>=8.3.176",
    "uvicorn>=0.35.0",
    "whitenoise>=6.9.0",
]
